+++ src/frontend/web_application/src/services/importance-level/index.js
@@ -1,0 +1,6 @@
+export const IL_MIN = -10;
+export const IL_MAX = 10;
+
+export const importanceLevelHeader = {
+  'X-Caliopen-IL': `${IL_MIN};${IL_MAX}`,
+};