+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Connect settings and apply
 - The brand new Timeline
 - Load more in the discussions
+- API for importance level messages filtering
+- API for full-text searches on messages & contacts
+- Improve `PATCH` API
+- Backend produces `excerpt` for messages.
 
 ## [0.3.0] 2017-08-31